fix 23440
