+++ test/extended/operators/olm.go
@@ -1,19 +1,14 @@
 package operators
 
 import (
-	"encoding/json"
 	"fmt"
 	"regexp"
 
 	g "github.com/onsi/ginkgo"
 	o "github.com/onsi/gomega"
 
-	"io/ioutil"
-	"math/rand"
 	"path/filepath"
-	"strconv"
 	"strings"
-	"sync"
 	"time"
 
 	exutil "github.com/openshift/openshift-tests/test/extended/util"
@@ -147,39 +142,36 @@ var _ = g.Describe("[sig-operators] OLM should", func() {
 var _ = g.Describe("[sig-operators] OLM for an end user use", func() {
 	defer g.GinkgoRecover()
 
-	var (
-		oc           = exutil.NewCLI("olm-23440", exutil.KubeConfigPath())
-		operatorWait = 120 * time.Second
+	var oc = exutil.NewCLI("olm", exutil.KubeConfigPath())
 
-		buildPruningBaseDir = exutil.FixturePath("testdata", "olm")
-		operatorGroup       = filepath.Join(buildPruningBaseDir, "operatorgroup.yaml")
-		etcdSub             = filepath.Join(buildPruningBaseDir, "etcd-subscription.yaml")
-		etcdCluster         = filepath.Join(buildPruningBaseDir, "etcd-cluster.yaml")
-	)
-
-	files := []string{operatorGroup, etcdSub}
 	// author: jiazha@redhat.com
 	g.It("Critical-23440-can subscribe to the etcd operator", func() {
-		g.By("Cluster-admin user subscribe the operator resource")
-		for _, v := range files {
-			configFile, err := oc.AsAdmin().Run("process").Args("--ignore-unknown-parameters=true", "-f", v, "-p", "NAME=test-operator", fmt.Sprintf("NAMESPACE=%s", oc.Namespace()), "SOURCENAME=community-operators", "SOURCENAMESPACE=openshift-marketplace").OutputToFile("config.json")
-			o.Expect(err).NotTo(o.HaveOccurred())
-			err = oc.AsAdmin().WithoutNamespace().Run("create").Args("-f", configFile).Execute()
-			o.Expect(err).NotTo(o.HaveOccurred())
+		buildPruningBaseDir := exutil.FixturePath("testdata", "olm")
+		etcdCluster := filepath.Join(buildPruningBaseDir, "etcd-cluster.yaml")
+		subTemplate := filepath.Join(buildPruningBaseDir, "olm-subscription.yaml")
+		oc.SetupProject()
 
+		dr := make(describerResrouce)
+		itName := g.CurrentGinkgoTestDescription().TestText
+		dr.addIr(itName)
+
+		g.By("Cluster-admin start to subscribe to etcd operator")
+		sub := subscriptionDescription{
+			subName:                "sub-23440",
+			namespace:              "openshift-operators",
+			catalogSourceName:      "community-operators",
+			catalogSourceNamespace: "openshift-marketplace",
+			channel:                "clusterwide-alpha",
+			ipApproval:             "Automatic",
+			operatorPackage:        "etcd",
+			singleNamespace:        false,
+			template:               subTemplate,
 		}
-		err := wait.Poll(10*time.Second, operatorWait, func() (bool, error) {
-			output, err := oc.AsAdmin().Run("get").Args("-n", oc.Namespace(), "csv", "etcdoperator.v0.9.4", "-o=jsonpath={.status.phase}").Output()
-			if err != nil {
-				e2e.Failf("Failed to deploy etcdoperator.v0.9.4, error:%v", err)
-				return false, err
-			}
-			if strings.Contains(output, "Succeeded") {
-				return true, nil
-			}
-			return false, nil
-		})
-		o.Expect(err).NotTo(o.HaveOccurred())
+		sub.create(oc, itName, dr)
+		defer sub.delete(itName, dr)
+		defer sub.getCSV().delete(itName, dr)
+		newCheck("expect", asAdmin, true, compare, "Succeeded", ok, []string{"csv", sub.installedCSV, "-o=jsonpath={.status.phase}"}).check(oc)
+
 
 		g.By("Switch to common user to create the resources provided by the operator")
 		etcdClusterName := "example-etcd-cluster"
@@ -188,23 +180,14 @@ var _ = g.Describe("[sig-operators] OLM for an end user use", func() {
 		err = oc.Run("create").Args("-f", configFile).Execute()
 		o.Expect(err).NotTo(o.HaveOccurred())
 
-		err = wait.Poll(10*time.Second, operatorWait, func() (bool, error) {
-			output, err := oc.Run("get").Args("-n", oc.Namespace(), "etcdCluster", etcdClusterName, "-o=jsonpath={.status}").Output()
-			if err != nil {
-				e2e.Failf("Failed to get etcdCluster, error:%v", err)
-				return false, err
-			}
-			if strings.Contains(output, "phase:Running") && strings.Contains(output, "currentVersion:3.2.13") && strings.Contains(output, "size:3") {
-				return true, nil
-			}
-			return false, nil
-		})
-		o.Expect(err).NotTo(o.HaveOccurred())
-		output, err := oc.Run("get").Args("pods", "-n", oc.Namespace()).Output()
-		o.Expect(err).NotTo(o.HaveOccurred())
-		o.Expect(output).To(o.ContainSubstring(etcdClusterName))
-	})
+		defer func() {
+			_, err := oc.Run("delete").Args("etcdcluster", etcdClusterName, "-n", oc.Namespace()).Output()
+			o.Expect(err).NotTo(o.HaveOccurred())
+		}()
 
+		newCheck("expect", false, true, compare, "Running", ok, []string{"etcdCluster", etcdClusterName, "-o=jsonpath={.status.phase}"}).check(oc)
+
+	})
 })
 
 var _ = g.Describe("[sig-operators] OLM for an end user handle within a namespace", func() {
@@ -224,11 +207,11 @@ var _ = g.Describe("[sig-operators] OLM for an end user handle within a namespac
 			template:  ogSingleTemplate,
 		}
 		subD = subscriptionDescription{
-			name:                   "hawtio-operator",
+			subName:                "hawtio-operator",
 			namespace:              "",
 			channel:                "alpha",
 			ipApproval:             "Automatic",
-			operator:               "hawtio-operator",
+			operatorPackage:        "hawtio-operator",
 			catalogSourceName:      "community-operators",
 			catalogSourceNamespace: "openshift-marketplace",
 			startingCSV:            "",
@@ -335,11 +318,11 @@ var _ = g.Describe("[sig-operators] OLM for an end user handle to support", func
 			template:    catsrcCmTemplate,
 		}
 		subNc = subscriptionDescription{
-			name:                   "namespace-configuration-operator",
+			subName:                "namespace-configuration-operator",
 			namespace:              "", //must be set in iT
 			channel:                "alpha",
 			ipApproval:             "Automatic",
-			operator:               "namespace-configuration-operator",
+			operatorPackage:        "namespace-configuration-operator",
 			catalogSourceName:      "catsrc-community-namespaceconfig-operators",
 			catalogSourceNamespace: "", //must be set in iT
 			startingCSV:            "",
@@ -475,11 +458,11 @@ var _ = g.Describe("[sig-operators] OLM for an end user handle within all namesp
 		var (
 			itName = g.CurrentGinkgoTestDescription().TestText
 			sub    = subscriptionDescription{
-				name:                   "teiid",
+				subName:                "teiid",
 				namespace:              "openshift-operators",
 				channel:                "beta",
 				ipApproval:             "Automatic",
-				operator:               "teiid",
+				operatorPackage:        "teiid",
 				catalogSourceName:      "community-operators",
 				catalogSourceNamespace: "openshift-marketplace",
 				// startingCSV:            "teiid.v0.3.0",
@@ -546,11 +529,11 @@ var _ = g.Describe("[sig-operators] OLM for an end user handle within all namesp
 		var (
 			itName           = g.CurrentGinkgoTestDescription().TestText
 			subElasticSearch = subscriptionDescription{
-				name:                   "elasticsearch-operator",
+				subName:                "elasticsearch-operator",
 				namespace:              "openshift-operators",
 				channel:                "preview",
 				ipApproval:             "Automatic",
-				operator:               "elasticsearch-operator",
+				operatorPackage:        "elasticsearch-operator",
 				catalogSourceName:      "redhat-operators",
 				catalogSourceNamespace: "openshift-marketplace",
 				// startingCSV:            "elasticsearch-operator.4.1.37-202003021622",
@@ -568,11 +551,11 @@ var _ = g.Describe("[sig-operators] OLM for an end user handle within all namesp
 			}
 
 			subJaeger = subscriptionDescription{
-				name:                   "jaeger-product",
+				subName:                "jaeger-product",
 				namespace:              "openshift-operators",
 				channel:                "stable",
 				ipApproval:             "Automatic",
-				operator:               "jaeger-product",
+				operatorPackage:        "jaeger-product",
 				catalogSourceName:      "redhat-operators",
 				catalogSourceNamespace: "openshift-marketplace",
 				// startingCSV:            "jaeger-operator.v1.17.1",
@@ -632,763 +615,3 @@ var _ = g.Describe("[sig-operators] OLM for an end user handle within all namesp
 	})
 
 })
-
-const (
-	asAdmin          = true
-	asUser           = false
-	withoutNamespace = true
-	withNamespace    = false
-	compare          = true
-	contain          = false
-	requireNS        = true
-	notRequireNS     = false
-	present          = true
-	notPresent       = false
-	ok               = true
-	nok              = false
-)
-
-type csvDescription struct {
-	name      string
-	namespace string
-}
-
-func (csv csvDescription) delete(itName string, dr describerResrouce) {
-	dr.getIr(itName).remove(csv.name, "csv", csv.namespace)
-}
-
-type subscriptionDescription struct {
-	name                   string
-	namespace              string
-	channel                string
-	ipApproval             string
-	operator               string
-	catalogSourceName      string
-	catalogSourceNamespace string
-	startingCSV            string
-	currentCSV             string
-	installedCSV           string
-	template               string
-	singleNamespace        bool
-	ipCsv                  string
-}
-
-func (sub *subscriptionDescription) create(oc *exutil.CLI, itName string, dr describerResrouce) {
-	sub.createWithoutCheck(oc, itName, dr)
-	if strings.Compare(sub.ipApproval, "Automatic") == 0 {
-		sub.findInstalledCSV(oc, itName, dr)
-	} else {
-		newCheck("expect", asAdmin, withoutNamespace, compare, "UpgradePending", ok, []string{"sub", sub.name, "-n", sub.namespace, "-o=jsonpath={.status.state}"}).check(oc)
-	}
-}
-
-func (sub *subscriptionDescription) createWithoutCheck(oc *exutil.CLI, itName string, dr describerResrouce) {
-	isAutomatic := strings.Compare(sub.ipApproval, "Automatic") == 0
-	if strings.Compare(sub.currentCSV, "") == 0 {
-		sub.currentCSV = getResource(oc, asAdmin, withoutNamespace, "packagemanifest", sub.name, fmt.Sprintf("-o=jsonpath={.status.channels[?(@.name==\"%s\")].currentCSV}", sub.channel))
-		o.Expect(sub.currentCSV).NotTo(o.BeEmpty())
-	}
-	if isAutomatic {
-		sub.startingCSV = sub.currentCSV
-	} else {
-		o.Expect(sub.startingCSV).NotTo(o.BeEmpty())
-	}
-	err := applyResourceFromTemplate(oc, "--ignore-unknown-parameters=true", "-f", sub.template, "-p", "SUBNAME="+sub.name, "SUBNAMESPACE="+sub.namespace, "CHANNEL="+sub.channel,
-		"APPROVAL="+sub.ipApproval, "OPERATORNAME="+sub.operator, "SOURCENAME="+sub.catalogSourceName, "SOURCENAMESPACE="+sub.catalogSourceNamespace, "STARTINGCSV="+sub.startingCSV)
-	o.Expect(err).NotTo(o.HaveOccurred())
-	dr.getIr(itName).add(newResource(oc, "sub", sub.name, requireNS, sub.namespace))
-}
-
-func (sub *subscriptionDescription) findInstalledCSV(oc *exutil.CLI, itName string, dr describerResrouce) {
-	newCheck("expect", asAdmin, withoutNamespace, compare, "AtLatestKnown", ok, []string{"sub", sub.name, "-n", sub.namespace, "-o=jsonpath={.status.state}"}).check(oc)
-	installedCSV := getResource(oc, asAdmin, withoutNamespace, "sub", sub.name, "-n", sub.namespace, "-o=jsonpath={.status.installedCSV}")
-	o.Expect(installedCSV).NotTo(o.BeEmpty())
-	if strings.Compare(sub.installedCSV, installedCSV) != 0 {
-		sub.installedCSV = installedCSV
-		dr.getIr(itName).add(newResource(oc, "csv", sub.installedCSV, requireNS, sub.namespace))
-	}
-	e2e.Logf("the installed CSV name is %s", sub.installedCSV)
-}
-
-func (sub *subscriptionDescription) expectCSV(oc *exutil.CLI, itName string, dr describerResrouce, cv string) {
-	err := wait.Poll(3*time.Second, 180*time.Second, func() (bool, error) {
-		sub.findInstalledCSV(oc, itName, dr)
-		if strings.Compare(sub.installedCSV, cv) == 0 {
-			return true, nil
-		}
-		return false, nil
-	})
-	o.Expect(err).NotTo(o.HaveOccurred())
-}
-
-func (sub *subscriptionDescription) approve(oc *exutil.CLI, itName string, dr describerResrouce) {
-	err := wait.Poll(1*time.Second, 60*time.Second, func() (bool, error) {
-		for strings.Compare(sub.installedCSV, "") == 0 {
-			state := getResource(oc, asAdmin, withoutNamespace, "sub", sub.name, "-n", sub.namespace, "-o=jsonpath={.status.state}")
-			if strings.Compare(state, "AtLatestKnown") == 0 {
-				sub.installedCSV = getResource(oc, asAdmin, withoutNamespace, "sub", sub.name, "-n", sub.namespace, "-o=jsonpath={.status.installedCSV}")
-				dr.getIr(itName).add(newResource(oc, "csv", sub.installedCSV, requireNS, sub.namespace))
-				e2e.Logf("it is already done, and the installed CSV name is %s", sub.installedCSV)
-				continue
-			}
-
-			ipCsv := getResource(oc, asAdmin, withoutNamespace, "sub", sub.name, "-n", sub.namespace, "-o=jsonpath={.status.installplan.name}{\" \"}{.status.currentCSV}")
-			sub.ipCsv = ipCsv + "##" + sub.ipCsv
-			installPlan := strings.Fields(ipCsv)[0]
-			o.Expect(installPlan).NotTo(o.BeEmpty())
-			e2e.Logf("try to approve installPlan %s", installPlan)
-			patchResource(oc, asAdmin, withoutNamespace, "ip", installPlan, "-n", sub.namespace, "--type", "merge", "-p", "{\"spec\": {\"approved\": true}}")
-			err := wait.Poll(3*time.Second, 10*time.Second, func() (bool, error) {
-				err := newCheck("expect", asAdmin, withoutNamespace, compare, "Complete", ok, []string{"ip", installPlan, "-n", sub.namespace, "-o=jsonpath={.status.phase}"}).checkWithoutAssert(oc)
-				if err != nil {
-					e2e.Logf("the get error is %v, and try next", err)
-					return false, nil
-				}
-				return true, nil
-			})
-			o.Expect(err).NotTo(o.HaveOccurred())
-		}
-		return true, nil
-	})
-	o.Expect(err).NotTo(o.HaveOccurred())
-}
-
-func (sub *subscriptionDescription) getCSV() csvDescription {
-	return csvDescription{sub.installedCSV, sub.namespace}
-}
-
-func (sub *subscriptionDescription) getInstanceVersion(oc *exutil.CLI) string {
-	version := ""
-	output := strings.Split(getResource(oc, asUser, withoutNamespace, "csv", sub.installedCSV, "-n", sub.namespace, "-o=jsonpath={.metadata.annotations.alm-examples}"), "\n")
-	for _, line := range output {
-		if strings.Contains(line, "\"version\"") {
-			version = strings.Trim(strings.Fields(strings.TrimSpace(line))[1], "\"")
-			break
-		}
-	}
-	o.Expect(version).NotTo(o.BeEmpty())
-	return version
-}
-
-func (sub *subscriptionDescription) createInstance(oc *exutil.CLI, instance string) {
-	path := filepath.Join(e2e.TestContext.OutputDir, sub.namespace+"-"+"instance.json")
-	err := ioutil.WriteFile(path, []byte(instance), 0644)
-	o.Expect(err).NotTo(o.HaveOccurred())
-	err = oc.AsAdmin().WithoutNamespace().Run("apply").Args("-n", sub.namespace, "-f", path).Execute()
-	o.Expect(err).NotTo(o.HaveOccurred())
-}
-
-func (sub *subscriptionDescription) delete(itName string, dr describerResrouce) {
-	dr.getIr(itName).remove(sub.name, "sub", sub.namespace)
-}
-
-func (sub *subscriptionDescription) patch(oc *exutil.CLI, patch string) {
-	patchResource(oc, asAdmin, withoutNamespace, "sub", sub.name, "-n", sub.namespace, "--type", "merge", "-p", patch)
-}
-
-type crdDescription struct {
-	name     string
-	template string
-}
-
-func (crd *crdDescription) create(oc *exutil.CLI, itName string, dr describerResrouce) {
-	err := applyResourceFromTemplate(oc, "--ignore-unknown-parameters=true", "-f", crd.template, "-p", "NAME="+crd.name)
-	o.Expect(err).NotTo(o.HaveOccurred())
-	dr.getIr(itName).add(newResource(oc, "crd", crd.name, notRequireNS, ""))
-}
-
-func (crd *crdDescription) delete(oc *exutil.CLI) {
-	removeResource(oc, asAdmin, withoutNamespace, "crd", crd.name)
-}
-
-type configMapDescription struct {
-	name      string
-	namespace string
-	template  string
-}
-
-func (cm *configMapDescription) create(oc *exutil.CLI, itName string, dr describerResrouce) {
-	err := applyResourceFromTemplate(oc, "--ignore-unknown-parameters=true", "-f", cm.template, "-p", "NAME="+cm.name, "NAMESPACE="+cm.namespace)
-	o.Expect(err).NotTo(o.HaveOccurred())
-	dr.getIr(itName).add(newResource(oc, "cm", cm.name, requireNS, cm.namespace))
-}
-func (cm *configMapDescription) patch(oc *exutil.CLI, patch string) {
-	patchResource(oc, asAdmin, withoutNamespace, "cm", cm.name, "-n", cm.namespace, "--type", "merge", "-p", patch)
-}
-
-type catalogSourceDescription struct {
-	name        string
-	namespace   string
-	displayName string
-	publisher   string
-	sourceType  string
-	address     string
-	template    string
-}
-
-func (catsrc *catalogSourceDescription) create(oc *exutil.CLI, itName string, dr describerResrouce) {
-	err := applyResourceFromTemplate(oc, "--ignore-unknown-parameters=true", "-f", catsrc.template,
-		"-p", "NAME="+catsrc.name, "NAMESPACE="+catsrc.namespace, "ADDRESS="+catsrc.address,
-		"DISPLAYNAME="+"\""+catsrc.displayName+"\"", "PUBLISHER="+"\""+catsrc.publisher+"\"", "SOURCETYPE="+catsrc.sourceType)
-	o.Expect(err).NotTo(o.HaveOccurred())
-	dr.getIr(itName).add(newResource(oc, "catsrc", catsrc.name, requireNS, catsrc.namespace))
-}
-func (catsrc *catalogSourceDescription) delete(itName string, dr describerResrouce) {
-	dr.getIr(itName).remove(catsrc.name, "catsrc", catsrc.namespace)
-}
-
-type operatorGroupDescription struct {
-	name         string
-	namespace    string
-	multinslabel string
-	template     string
-}
-
-func (og *operatorGroupDescription) create(oc *exutil.CLI, itName string, dr describerResrouce) {
-	var err error
-	if strings.Compare(og.multinslabel, "") == 0 {
-		err = applyResourceFromTemplate(oc, "--ignore-unknown-parameters=true", "-f", og.template, "-p", "NAME="+og.name, "NAMESPACE="+og.namespace)
-	} else {
-		err = applyResourceFromTemplate(oc, "--ignore-unknown-parameters=true", "-f", og.template, "-p", "NAME="+og.name, "NAMESPACE="+og.namespace, "MULTINSLABEL="+og.multinslabel)
-	}
-	o.Expect(err).NotTo(o.HaveOccurred())
-	dr.getIr(itName).add(newResource(oc, "og", og.name, requireNS, og.namespace))
-}
-func (og *operatorGroupDescription) delete(itName string, dr describerResrouce) {
-	dr.getIr(itName).remove(og.name, "og", og.namespace)
-}
-
-type operatorSourceDescription struct {
-	name              string
-	namespace         string
-	namelabel         string
-	registrynamespace string
-	displayname       string
-	publisher         string
-	template          string
-	deploymentName    string
-}
-
-func (osrc *operatorSourceDescription) create(oc *exutil.CLI, itName string, dr describerResrouce) {
-	err := applyResourceFromTemplate(oc, "--ignore-unknown-parameters=true", "-f", osrc.template, "-p", "NAME="+osrc.name, "NAMESPACE="+osrc.namespace,
-		"NAMELABEL="+osrc.namelabel, "REGISTRYNAMESPACE="+osrc.registrynamespace, "DISPLAYNAME="+osrc.displayname, "PUBLISHER="+osrc.publisher)
-	o.Expect(err).NotTo(o.HaveOccurred())
-	dr.getIr(itName).add(newResource(oc, "opsrc", osrc.name, requireNS, osrc.namespace))
-}
-
-func (osrc *operatorSourceDescription) delete(itName string, dr describerResrouce) {
-	dr.getIr(itName).remove(osrc.name, "opsrc", osrc.namespace)
-}
-
-func (osrc *operatorSourceDescription) getRunningNodes(oc *exutil.CLI) string {
-	nodesNames := getResource(oc, asAdmin, withoutNamespace, "pod", fmt.Sprintf("--selector=marketplace.operatorSource=%s", osrc.name), "-n", osrc.namespace, "-o=jsonpath={.items[*]..nodeName}")
-	o.Expect(nodesNames).NotTo(o.BeEmpty())
-	return nodesNames
-}
-func (osrc *operatorSourceDescription) getDeployment(oc *exutil.CLI) {
-	output := getResource(oc, asAdmin, withoutNamespace, "deployment", fmt.Sprintf("--selector=opsrc-owner-name=%s", osrc.name), "-n", osrc.namespace, "-o=jsonpath={.items[0].metadata.name}")
-	o.Expect(output).NotTo(o.BeEmpty())
-	osrc.deploymentName = output
-}
-func (osrc *operatorSourceDescription) patchDeployment(oc *exutil.CLI, content string) {
-	if strings.Compare(osrc.deploymentName, "") == 0 {
-		osrc.deploymentName = osrc.name
-	}
-	patchResource(oc, asAdmin, withoutNamespace, "deployment", osrc.deploymentName, "-n", osrc.namespace, "--type", "merge", "-p", content)
-}
-func (osrc *operatorSourceDescription) getTolerations(oc *exutil.CLI) string {
-	if strings.Compare(osrc.deploymentName, "") == 0 {
-		osrc.deploymentName = osrc.name
-	}
-	output := getResource(oc, asAdmin, withoutNamespace, "deployment", osrc.deploymentName, "-n", osrc.namespace, "-o=jsonpath={.spec.template.spec.tolerations}")
-	if strings.Compare(output, "") == 0 {
-		e2e.Logf("no tolerations %v", output)
-		return "\"tolerations\": null"
-	}
-	tolerations := "\"tolerations\": " + convertLMtoJSON(output)
-	e2e.Logf("the tolerations:===%v===", tolerations)
-	return tolerations
-}
-
-type catalogSourceConfigDescription struct {
-	name            string
-	namespace       string
-	packages        string
-	targetnamespace string
-	source          string
-	template        string
-}
-
-func (csc *catalogSourceConfigDescription) create(oc *exutil.CLI, itName string, dr describerResrouce) {
-	err := applyResourceFromTemplate(oc, "--ignore-unknown-parameters=true", "-f", csc.template, "-p", "NAME="+csc.name, "NAMESPACE="+csc.namespace,
-		"PACKAGES="+csc.packages, "TARGETNAMESPACE="+csc.targetnamespace, "SOURCE="+csc.source)
-	o.Expect(err).NotTo(o.HaveOccurred())
-	dr.getIr(itName).add(newResource(oc, "csc", csc.name, requireNS, csc.namespace))
-}
-
-func (csc *catalogSourceConfigDescription) delete(itName string, dr describerResrouce) {
-	dr.getIr(itName).remove(csc.name, "csc", csc.namespace)
-}
-
-type projectDescription struct {
-	name            string
-	targetNamespace string
-}
-
-func (p *projectDescription) create(oc *exutil.CLI, itName string, dr describerResrouce) {
-	removeResource(oc, asAdmin, withoutNamespace, "project", p.name)
-	_, err := doAction(oc, "new-project", asAdmin, withoutNamespace, p.name)
-	o.Expect(err).NotTo(o.HaveOccurred())
-	dr.getIr(itName).add(newResource(oc, "project", p.name, notRequireNS, ""))
-	_, err = doAction(oc, "project", asAdmin, withoutNamespace, p.targetNamespace)
-	o.Expect(err).NotTo(o.HaveOccurred())
-}
-
-func (p *projectDescription) label(oc *exutil.CLI, label string) {
-	_, err := doAction(oc, "label", asAdmin, withoutNamespace, "ns", p.name, "env="+label)
-	o.Expect(err).NotTo(o.HaveOccurred())
-}
-
-func (p *projectDescription) delete(oc *exutil.CLI) {
-	_, err := doAction(oc, "delete", asAdmin, withoutNamespace, "project", p.name)
-	o.Expect(err).NotTo(o.HaveOccurred())
-}
-
-type serviceAccountDescription struct {
-	name           string
-	namespace      string
-	definitionfile string
-}
-
-func newSa(name, namespace string) *serviceAccountDescription {
-	return &serviceAccountDescription{
-		name:           name,
-		namespace:      namespace,
-		definitionfile: "",
-	}
-}
-func (sa *serviceAccountDescription) getDefinition(oc *exutil.CLI) {
-	parameters := []string{"sa", sa.name, "-n", sa.namespace, "-o=json"}
-	definitionfile, err := oc.AsAdmin().WithoutNamespace().Run("get").Args(parameters...).OutputToFile("sa-config.json")
-	o.Expect(err).NotTo(o.HaveOccurred())
-	sa.definitionfile = definitionfile
-}
-func (sa *serviceAccountDescription) delete(oc *exutil.CLI) {
-	_, err := doAction(oc, "delete", asAdmin, withoutNamespace, "sa", sa.name, "-n", sa.namespace)
-	o.Expect(err).NotTo(o.HaveOccurred())
-}
-func (sa *serviceAccountDescription) reapply(oc *exutil.CLI) {
-	err := oc.AsAdmin().WithoutNamespace().Run("apply").Args("-f", sa.definitionfile).Execute()
-	o.Expect(err).NotTo(o.HaveOccurred())
-}
-func (sa *serviceAccountDescription) checkAuth(oc *exutil.CLI, expected string, cr string) {
-	err := wait.Poll(3*time.Second, 150*time.Second, func() (bool, error) {
-		output, _ := doAction(oc, "auth", asAdmin, withNamespace, "--as", fmt.Sprintf("system:serviceaccount:%s:%s", sa.namespace, sa.name), "can-i", "create", cr)
-		e2e.Logf("the result of checkAuth:%v", output)
-		if strings.Contains(output, expected) {
-			return true, nil
-		}
-		return false, nil
-	})
-	o.Expect(err).NotTo(o.HaveOccurred())
-}
-
-type roleDescription struct {
-	name      string
-	namespace string
-}
-
-func newRole(name string, namespace string) *roleDescription {
-	return &roleDescription{
-		name:      name,
-		namespace: namespace,
-	}
-}
-func (role *roleDescription) patch(oc *exutil.CLI, patch string) {
-	patchResource(oc, asAdmin, withoutNamespace, "role", role.name, "-n", role.namespace, "--type", "merge", "-p", patch)
-}
-func (role *roleDescription) getRules(oc *exutil.CLI) string {
-	return role.getRulesWithDelete(oc, "nodelete")
-}
-func (role *roleDescription) getRulesWithDelete(oc *exutil.CLI, delete string) string {
-	var roleboday map[string]interface{}
-	output := getResource(oc, asAdmin, withoutNamespace, "role", role.name, "-n", role.namespace, "-o=json")
-	err := json.Unmarshal([]byte(output), &roleboday)
-	o.Expect(err).NotTo(o.HaveOccurred())
-	rules := roleboday["rules"].([]interface{})
-
-	handleRuleAttribute := func(rc *strings.Builder, rt string, r map[string]interface{}) {
-		rc.WriteString("\"" + rt + "\":[")
-		items := r[rt].([]interface{})
-		e2e.Logf("%s:%v, and the len:%v", rt, items, len(items))
-		for i, v := range items {
-			vc := v.(string)
-			rc.WriteString("\"" + vc + "\"")
-			if i != len(items)-1 {
-				rc.WriteString(",")
-			}
-		}
-		rc.WriteString("]")
-		if strings.Compare(rt, "verbs") != 0 {
-			rc.WriteString(",")
-		}
-	}
-
-	var rc strings.Builder
-	rc.WriteString("[")
-	for _, rv := range rules {
-		rule := rv.(map[string]interface{})
-		if strings.Compare(delete, "nodelete") != 0 && strings.Compare(rule["apiGroups"].([]interface{})[0].(string), delete) == 0 {
-			continue
-		}
-
-		rc.WriteString("{")
-		handleRuleAttribute(&rc, "apiGroups", rule)
-		handleRuleAttribute(&rc, "resources", rule)
-		handleRuleAttribute(&rc, "verbs", rule)
-		rc.WriteString("},")
-	}
-	result := strings.TrimSuffix(rc.String(), ",") + "]"
-	e2e.Logf("rc:%v", result)
-	return result
-}
-
-type checkDescription struct {
-	method          string
-	executor        bool
-	inlineNamespace bool
-	expectAction    bool
-	expectContent   string
-	expect          bool
-	resource        []string
-}
-
-func newCheck(method string, executor bool, inlineNamespace bool, expectAction bool,
-	expectContent string, expect bool, resource []string) checkDescription {
-	return checkDescription{
-		method:          method,
-		executor:        executor,
-		inlineNamespace: inlineNamespace,
-		expectAction:    expectAction,
-		expectContent:   expectContent,
-		expect:          expect,
-		resource:        resource,
-	}
-}
-func (ck checkDescription) check(oc *exutil.CLI) {
-	switch ck.method {
-	case "present":
-		ok := isPresentResource(oc, ck.executor, ck.inlineNamespace, ck.expectAction, ck.resource...)
-		o.Expect(ok).To(o.BeTrue())
-	case "expect":
-		err := expectedResource(oc, ck.executor, ck.inlineNamespace, ck.expectAction, ck.expectContent, ck.expect, ck.resource...)
-		o.Expect(err).NotTo(o.HaveOccurred())
-	default:
-		err := fmt.Errorf("unknown method")
-		o.Expect(err).NotTo(o.HaveOccurred())
-	}
-}
-func (ck checkDescription) checkWithoutAssert(oc *exutil.CLI) error {
-	switch ck.method {
-	case "present":
-		ok := isPresentResource(oc, ck.executor, ck.inlineNamespace, ck.expectAction, ck.resource...)
-		if ok {
-			return nil
-		}
-		return fmt.Errorf("it is not epxected")
-	case "expect":
-		return expectedResource(oc, ck.executor, ck.inlineNamespace, ck.expectAction, ck.expectContent, ck.expect, ck.resource...)
-	default:
-		return fmt.Errorf("unknown method")
-	}
-}
-
-type checkList []checkDescription
-
-func (cl checkList) add(ck checkDescription) {
-	cl = append(cl, ck)
-}
-func (cl checkList) empty() {
-	cl = cl[0:0]
-}
-func (cl checkList) check(oc *exutil.CLI) {
-	var wg sync.WaitGroup
-	for _, ck := range cl {
-		wg.Add(1)
-		go func(ck checkDescription) {
-			defer g.GinkgoRecover()
-			defer wg.Done()
-			ck.check(oc)
-		}(ck)
-	}
-	wg.Wait()
-}
-
-type resourceDescription struct {
-	oc               *exutil.CLI
-	asAdmin          bool
-	withoutNamespace bool
-	kind             string
-	name             string
-	requireNS        bool
-	namespace        string
-}
-
-func newResource(oc *exutil.CLI, kind string, name string, nsflag bool, namespace string) resourceDescription {
-	return resourceDescription{
-		oc:               oc,
-		asAdmin:          asAdmin,
-		withoutNamespace: withoutNamespace,
-		kind:             kind,
-		name:             name,
-		requireNS:        nsflag,
-		namespace:        namespace,
-	}
-}
-func (r resourceDescription) delete() {
-	if r.withoutNamespace && r.requireNS {
-		removeResource(r.oc, r.asAdmin, r.withoutNamespace, r.kind, r.name, "-n", r.namespace)
-	} else {
-		removeResource(r.oc, r.asAdmin, r.withoutNamespace, r.kind, r.name)
-	}
-}
-
-type itResource map[string]resourceDescription
-
-func (ir itResource) add(r resourceDescription) {
-	ir[r.name+r.kind+r.namespace] = r
-}
-func (ir itResource) get(name string, kind string, namespace string) resourceDescription {
-	r, ok := ir[name+kind+namespace]
-	o.Expect(ok).To(o.BeTrue())
-	return r
-}
-func (ir itResource) remove(name string, kind string, namespace string) {
-	rKey := name + kind + namespace
-	if r, ok := ir[rKey]; ok {
-		r.delete()
-		delete(ir, rKey)
-	}
-}
-func (ir itResource) cleanup() {
-	for _, r := range ir {
-		e2e.Logf("cleanup resource %s,   %s", r.kind, r.name)
-		ir.remove(r.name, r.kind, r.namespace)
-	}
-}
-
-type describerResrouce map[string]itResource
-
-func (dr describerResrouce) addIr(itName string) {
-	dr[itName] = itResource{}
-}
-func (dr describerResrouce) getIr(itName string) itResource {
-	ir, ok := dr[itName]
-	o.Expect(ok).To(o.BeTrue())
-	return ir
-}
-func (dr describerResrouce) rmIr(itName string) {
-	delete(dr, itName)
-}
-
-func convertLMtoJSON(content string) string {
-	var jb strings.Builder
-	jb.WriteString("[")
-	items := strings.Split(strings.TrimSuffix(strings.TrimPrefix(content, "["), "]"), "map")
-	for _, item := range items {
-		if strings.Compare(item, "") == 0 {
-			continue
-		}
-		kvs := strings.Fields(strings.TrimSuffix(strings.TrimPrefix(item, "["), "]"))
-		jb.WriteString("{")
-		for ki, kv := range kvs {
-			p := strings.Split(kv, ":")
-			jb.WriteString("\"" + p[0] + "\":")
-			jb.WriteString("\"" + p[1] + "\"")
-			if ki < len(kvs)-1 {
-				jb.WriteString(", ")
-			}
-		}
-		jb.WriteString("},")
-	}
-	return strings.TrimSuffix(jb.String(), ",") + "]"
-}
-
-func getRandomString() string {
-	chars := "abcdefghijklmnopqrstuvwxyz0123456789"
-	seed := rand.New(rand.NewSource(time.Now().UnixNano()))
-	buffer := make([]byte, 8)
-	for index := range buffer {
-		buffer[index] = chars[seed.Intn(len(chars))]
-	}
-	return string(buffer)
-}
-
-func generateUpdatedKubernatesVersion(oc *exutil.CLI) string {
-	subKubeVersions := strings.Split(getKubernetesVersion(oc), ".")
-	zVersion, _ := strconv.Atoi(subKubeVersions[1])
-	subKubeVersions[1] = strconv.Itoa(zVersion + 1)
-	return strings.Join(subKubeVersions[0:2], ".") + ".0"
-}
-
-func getKubernetesVersion(oc *exutil.CLI) string {
-	output, err := doAction(oc, "version", asAdmin, withoutNamespace, "-o=json")
-	o.Expect(err).NotTo(o.HaveOccurred())
-
-	var result map[string]interface{}
-	err = json.Unmarshal([]byte(output), &result)
-	o.Expect(err).NotTo(o.HaveOccurred())
-
-	gitVersion := result["serverVersion"].(map[string]interface{})["gitVersion"]
-	e2e.Logf("gitVersion is %v", gitVersion)
-	return strings.TrimPrefix(gitVersion.(string), "v")
-}
-
-func applyResourceFromTemplate(oc *exutil.CLI, parameters ...string) error {
-	var configFile string
-	err := wait.Poll(3*time.Second, 15*time.Second, func() (bool, error) {
-		output, err := oc.AsAdmin().Run("process").Args(parameters...).OutputToFile(getRandomString() + "olm-config.json")
-		if err != nil {
-			e2e.Logf("the err:%v, and try next round", err)
-			return false, nil
-		}
-		configFile = output
-		return true, nil
-	})
-	o.Expect(err).NotTo(o.HaveOccurred())
-
-	e2e.Logf("the file of resource is %s", configFile)
-	return oc.AsAdmin().WithoutNamespace().Run("apply").Args("-f", configFile).Execute()
-}
-
-func isPresentResource(oc *exutil.CLI, asAdmin bool, withoutNamespace bool, present bool, parameters ...string) bool {
-	parameters = append(parameters, "--ignore-not-found")
-	err := wait.Poll(3*time.Second, 60*time.Second, func() (bool, error) {
-		output, err := doAction(oc, "get", asAdmin, withoutNamespace, parameters...)
-		if err != nil {
-			e2e.Logf("the get error is %v, and try next", err)
-			return false, nil
-		}
-		if !present && strings.Compare(output, "") == 0 {
-			return true, nil
-		}
-		if present && strings.Compare(output, "") != 0 {
-			return true, nil
-		}
-		return false, nil
-	})
-	if err != nil {
-		return false
-	}
-	return true
-}
-
-func patchResource(oc *exutil.CLI, asAdmin bool, withoutNamespace bool, parameters ...string) {
-	_, err := doAction(oc, "patch", asAdmin, withoutNamespace, parameters...)
-	o.Expect(err).NotTo(o.HaveOccurred())
-}
-
-func execResource(oc *exutil.CLI, asAdmin bool, withoutNamespace bool, parameters ...string) string {
-	var result string
-	err := wait.Poll(3*time.Second, 6*time.Second, func() (bool, error) {
-		output, err := doAction(oc, "exec", asAdmin, withoutNamespace, parameters...)
-		if err != nil {
-			e2e.Logf("the exec error is %v, and try next", err)
-			return false, nil
-		}
-		result = output
-		return true, nil
-	})
-	o.Expect(err).NotTo(o.HaveOccurred())
-	e2e.Logf("the result of exec resource:%v", result)
-	return result
-}
-
-func getResource(oc *exutil.CLI, asAdmin bool, withoutNamespace bool, parameters ...string) string {
-	var result string
-	err := wait.Poll(3*time.Second, 120*time.Second, func() (bool, error) {
-		output, err := doAction(oc, "get", asAdmin, withoutNamespace, parameters...)
-		if err != nil {
-			e2e.Logf("the get error is %v, and try next", err)
-			return false, nil
-		}
-		result = output
-		return true, nil
-	})
-	o.Expect(err).NotTo(o.HaveOccurred())
-	e2e.Logf("the result of queried resource:%v", result)
-	return result
-}
-
-func expectedResource(oc *exutil.CLI, asAdmin bool, withoutNamespace bool, isCompare bool, content string, expect bool, parameters ...string) error {
-	cc := func(a, b string, ic bool) bool {
-		bs := strings.Split(b, "+2+")
-		ret := false
-		for _, s := range bs {
-			if (ic && strings.Compare(a, s) == 0) || (!ic && strings.Contains(a, s)) {
-				ret = true
-			}
-		}
-		return ret
-	}
-	return wait.Poll(3*time.Second, 150*time.Second, func() (bool, error) {
-		output, err := doAction(oc, "get", asAdmin, withoutNamespace, parameters...)
-		if err != nil {
-			e2e.Logf("the get error is %v, and try next", err)
-			return false, nil
-		}
-		e2e.Logf("the queried resource:%s", output)
-		if isCompare && expect && cc(output, content, isCompare) {
-			e2e.Logf("the output %s matches one of the content %s, expected", output, content)
-			return true, nil
-		}
-		if isCompare && !expect && !cc(output, content, isCompare) {
-			e2e.Logf("the output %s does not matche the content %s, expected", output, content)
-			return true, nil
-		}
-		if !isCompare && expect && cc(output, content, isCompare) {
-			e2e.Logf("the output %s contains one of the content %s, expected", output, content)
-			return true, nil
-		}
-		if !isCompare && !expect && !cc(output, content, isCompare) {
-			e2e.Logf("the output %s does not contain the content %s, expected", output, content)
-			return true, nil
-		}
-		return false, nil
-	})
-}
-
-func removeResource(oc *exutil.CLI, asAdmin bool, withoutNamespace bool, parameters ...string) {
-	output, err := doAction(oc, "delete", asAdmin, withoutNamespace, parameters...)
-	if err != nil && (strings.Contains(output, "NotFound") || strings.Contains(output, "No resources found")) {
-		e2e.Logf("the resource is deleted already")
-		return
-	}
-	o.Expect(err).NotTo(o.HaveOccurred())
-
-	err = wait.Poll(3*time.Second, 120*time.Second, func() (bool, error) {
-		output, err := doAction(oc, "get", asAdmin, withoutNamespace, parameters...)
-		if err != nil && (strings.Contains(output, "NotFound") || strings.Contains(output, "No resources found")) {
-			e2e.Logf("the resource is delete successfully")
-			return true, nil
-		}
-		return false, nil
-	})
-	o.Expect(err).NotTo(o.HaveOccurred())
-}
-
-func doAction(oc *exutil.CLI, action string, asAdmin bool, withoutNamespace bool, parameters ...string) (string, error) {
-	if asAdmin && withoutNamespace {
-		return oc.AsAdmin().WithoutNamespace().Run(action).Args(parameters...).Output()
-	}
-	if asAdmin && !withoutNamespace {
-		return oc.AsAdmin().Run(action).Args(parameters...).Output()
-	}
-	if !asAdmin && withoutNamespace {
-		return oc.WithoutNamespace().Run(action).Args(parameters...).Output()
-	}
-	if !asAdmin && !withoutNamespace {
-		return oc.Run(action).Args(parameters...).Output()
-	}
-	return "", nil
-}
