+++ test/extended/testdata/olm/etcd-cluster.yaml
@@ -6,9 +6,12 @@ objects:
 - apiVersion: etcd.database.coreos.com/v1beta2
   kind: EtcdCluster
   metadata:
+    annotations:
+      etcd.database.coreos.com/scope: clusterwide
     name: "${NAME}"
     namespace: "${NAMESPACE}"
   spec:
+    repository: quay.io/coreos/etcd
     size: 3
     version: 3.2.13
 parameters:
